* fixed a blunder with lights not stretching. * (hopefully) a working x11 clipboard (requires gtk and gdk)

+++ trunk/engine.h
@@ -55,7 +55,7 @@ class Engine
 
 		virtual bool isSupported() = 0;
 
-		virtual void drawLightmap(vertex* lightmap, int type, int width, int height, int scale);
+		virtual void drawLightmap(vertex* lightmap, int type, int width, int height, float scale);
 		virtual void doResize(int& w, int& h);
 		virtual void drawRectangle(float x, float y, float width, float height, oRGBA color) = 0;
 		virtual void drawRectangleLines(float x, float y, float width, float height, oRGBA color, float thickness = 1.f) {}
